Blocked Weighted Bootstrap - vectorised and parallel

This set of functions is an alternative to the bootBW() function. This set attempts to make the blocked weighted bootstrap algorithm more efficient through vectorisation and use of parallelisation techniques. The function syntax has been kept consistent with bootBW() for ease of transition. A more in depth discussion of the efficiencies gained from this alternative function is discussed here.

Usage

boot_bw(
  x,
  w,
  statistic,
  params,
  outputColumns = params,
  replicates = 400,
  cores = parallelly::availableCores(omit = 1)
)

boot_bw_weight(w)

boot_bw_sample_clusters(x, df_weighted, index = FALSE)

boot_bw_sample_within_clusters(cluster_df)

Arguments

x

A data.frame() with primary sampling unit (PSU) in variable named psu and at least one other variable containing data for estimation.

w

A data.frame() with primary sampling unit (PSU) in variable named psu and survey weights (i.e. PSU population) in variable named pop.

statistic

Am estimator function operating on variables in x containing data for estimation. The functions bootClassic() and bootPROBIT() are examples.

params

Parameters specified as names of columns in x that are to be passed to the function specified in statistic.

outputColumns

Names to be used for columns in output data.frame(). Default to names specified in params.

replicates

Number of bootstrap replicates to be performed. Default is 400.

cores

The number of computer cores to use or number of child processes to be run simultaneously. Default to one less than the available number of cores on current machine.

df_weighted

A data.frame() based on w with additional variables for weight and cumWeight. This is usually produced using the function boot_bw_weight().

index

Logical. Should index values be returned or a list of data.frame()s. Default to FALSE.

cluster_df

A list of data.frame()s for selected clusters.

Value

For boot_bw(), a data.frame() with number of columns equal to length of outputColumns; number of rows equal to number of replicates; and, names of variables equal to values of outputColumns. For boot_bw_weight(), A data.frame() based on w with two additional variables for weight and cumWeight. For boot_bw_sample_clusters(), either a vector of integers corresponding to the primary sampling unit (psu) identifier of the selected clusters (when index = TRUE) or a list of data.frame()s corresponding to the data for the selected clusters (when index = FALSE). For boot_bw_sample_within_clusters(), a matrix similar in structure to x of resampled data from each selected cluster.

Examples

boot_bw(
  x = indicatorsHH, w = villageData, statistic = bootClassic, 
  params = "anc1", replicates = 49
)
#> ✔ `w` has the needed variables with the appropriate names
#> ✔ `x` has the appropriate/expected data structure
#> ✔ "anc1" is a variable in `x`
#>         anc1
#> 1  0.2497630
#> 2  0.2217494
#> 3  0.2111316
#> 4  0.2635883
#> 5  0.2847587
#> 6  0.2007630
#> 7  0.2856471
#> 8  0.2478219
#> 9  0.2399038
#> 10 0.2474032
#> 11 0.2286785
#> 12 0.2235012
#> 13 0.2333011
#> 14 0.1856459
#> 15 0.2203844
#> 16 0.2469018
#> 17 0.2114286
#> 18 0.2320789
#> 19 0.1939163
#> 20 0.2775097
#> 21 0.2715455
#> 22 0.2597712
#> 23 0.1778412
#> 24 0.1903164
#> 25 0.2188801
#> 26 0.2714561
#> 27 0.2389094
#> 28 0.2284523
#> 29 0.2285714
#> 30 0.2005636
#> 31 0.2633287
#> 32 0.2470643
#> 33 0.2034968
#> 34 0.2358354
#> 35 0.2325792
#> 36 0.2214485
#> 37 0.2737244
#> 38 0.2223278
#> 39 0.2154356
#> 40 0.2292264
#> 41 0.2237796
#> 42 0.2199513
#> 43 0.2215281
#> 44 0.2285847
#> 45 0.2398148
#> 46 0.1870192
#> 47 0.2130150
#> 48 0.2274927
#> 49 0.2434944