Child Morbidity, Health Service Coverage, Anthropometry

Child indicators on morbidity, health service coverage and anthropometry calculated from survey data collected in survey conducted in 4 districts from 3 regions in Somalia.

indicatorsCH1

Format

A data frame with 14 columns and 3090 rows.

psu

The PSU identifier. This must use the same coding system used to identify the PSUs that is used in the indicators dataset

mID

The mother identifier

cID

The child identifier

ch1

Diarrhoea in the past 2 weeks (0/1)

ch2

Fever in the past 2 weeks (0/1)

ch3

Cough in the past 2 weeks (0/1)

ch4

Immunisation card (0/1)

ch5

BCG immunisation (0/1)

ch6

Vitamin A coverage in the past month (0/1)

ch7

Anti-helminth coverage in the past month (0/1)

sex

Sex of child

muac1

Mid-upper arm circumference in mm

muac2

Mid-upper arm circumference in mm

oedema

Oedema (0/1)

Source

Mother and child health and nutrition survey in 3 regions of Somalia