Mother Indicators Dataset

Mother indicators for health and nutrition calculated from survey data collected in survey conducted in 4 districts from 3 regions in Somalia.

indicatorsHH

Format

A data frame with 24 columns and 2136 rows:

psu

The PSU identifier. This must use the same coding system used to identify the PSUs that is used in the indicators dataset

mID

The mother identifier

mMUAC

Mothers with mid-upper arm circumference < 230 mm (0/1)

anc1

At least 1 antenatal care visit with a trained health professional (0/1)

anc2

At least 4 antenatal care visits with any service provider (0/1)

anc3

FeFol coverage (0/1)

anc4

Vitamin A coverage (0/1)

wash1

Improved sources of drinking water (0/1)

wash2

Improved sources of other water (0/1)

wash3

Probable safe drinking water (0/1)

wash4

Number of litres of water collected in a day

wash5

Improved toilet facilities (0/1)

wash6

Human waste disposal practices / behaviour (0/1)

wash7a

Handwashing score (from 0 to 5)

wash7b

Handwashing score of 5 (0/1)

hhs1

Household hunger score (from 0 to 6)

hhs2

Little or no hunger (0/1)

hhs3

Moderate hunger (0/1)

hhs4

Severe hunger (0/1)

mfg

Mother's dietary diversity score

pVitA

Plant-based vitamin A-rich foods (0/1)

aVitA

Animal-based vitamin A-rich foods (0/1)

xVitA

Any vitamin A-rich foods (0/1)

iron

Iron-rich foods (0/1)

Source

Mother and child health and nutrition survey in 3 regions of Somalia