Blocked Weighted Bootstrap - vectorised and parallel

This set of functions is an alternative to the bootBW() function. This set attempts to make the blocked weighted bootstrap algorithm more efficient through vectorisation and use of parallelisation techniques. The function syntax has been kept consistent with bootBW() for ease of transition. A more in depth discussion of the efficiencies gained from this alternative function is discussed here.

Usage

boot_bw(
  x,
  w,
  statistic,
  params,
  outputColumns = params,
  replicates = 400,
  strata = NULL,
  parallel = FALSE,
  cores = parallelly::availableCores(omit = 1)
)

boot_bw_parallel(
  x,
  w,
  statistic,
  params,
  outputColumns = params,
  replicates = 400,
  strata = NULL,
  cores = parallelly::availableCores(omit = 1)
)

boot_bw_sequential(
  x,
  w,
  statistic,
  params,
  outputColumns = params,
  replicates = 400,
  strata = NULL
)

boot_bw_weight(w)

boot_bw_sample_clusters(x, w, index = FALSE)

boot_bw_sample_within_clusters(cluster_df)

Arguments

x

A data.frame() with primary sampling unit (PSU) in variable named psu and at least one other variable containing data for estimation.

w

A data.frame() with primary sampling unit (PSU) in variable named psu and survey weights (i.e. PSU population) in variable named pop.

statistic

Am estimator function operating on variables in x containing data for estimation. The functions bootClassic() and bootPROBIT() are examples.

params

Parameters specified as names of columns in x that are to be passed to the function specified in statistic.

outputColumns

Names to be used for columns in output data.frame(). Default to names specified in params.

replicates

Number of bootstrap replicates to be performed. Default is 400.

strata

A character value for name of variable in x providing information on how x is grouped such that resampling is performed for each group. Default to NULL for no grouping and resampling is performed for full data.

parallel

Logical. Should resampling be done in parallel? Default to FALSE.

cores

The number of computer cores to use or number of child processes to be run simultaneously. Default to one less than the available number of cores on current machine.

index

Logical. Should index values be returned or a list of data.frame()s. Default to FALSE.

cluster_df

A list of data.frame()s for selected clusters.

Value

For boot_bw(), a data.frame() with number of columns equal to length of outputColumns; number of rows equal to number of replicates; and, names of variables equal to values of outputColumns. For boot_bw_weight(), A data.frame() based on w with two additional variables for weight and cumWeight. For boot_bw_sample_clusters(), either a vector of integers corresponding to the primary sampling unit (psu) identifier of the selected clusters (when index = TRUE) or a list of data.frame()s corresponding to the data for the selected clusters (when index = FALSE). For boot_bw_sample_within_clusters(), a matrix similar in structure to x of resampled data from each selected cluster.

Examples

boot_bw(
  x = indicatorsHH, w = villageData, statistic = bootClassic, 
  params = "anc1", replicates = 49, parallel = TRUE
)
#> 
#> ── Resampling in parallel ──
#> 
#> ℹ Setting up 3 parallel operations
#> ✔ Setting up 3 parallel operations [296ms]
#> 
#> ℹ Resampling with 49 replicates in parallel
#> ✔ Resampling with 49 replicates in parallel [2.7s]
#> 
#> ℹ Tidying up resampling outputs
#> ✔ Tidying up resampling outputs [26ms]
#> 
#> ℹ Closing 3 parallel operations
#> ✔ Closing 3 parallel operations [15ms]
#> 
#> $params
#> [1] "anc1"
#> 
#> $replicates
#> [1] 49
#> 
#> $strata
#> NULL
#> 
#> $boot_data
#>         anc1
#> 1  0.2497674
#> 2  0.2728558
#> 3  0.2172033
#> 4  0.2225300
#> 5  0.2372881
#> 6  0.1912331
#> 7  0.2076271
#> 8  0.2068302
#> 9  0.2516974
#> 10 0.2179427
#> 11 0.2349806
#> 12 0.2263875
#> 13 0.2038369
#> 14 0.2022308
#> 15 0.1796234
#> 16 0.1980337
#> 17 0.2060253
#> 18 0.2131072
#> 19 0.2227955
#> 20 0.2019774
#> 21 0.2116896
#> 22 0.2521531
#> 23 0.2874532
#> 24 0.2583616
#> 25 0.2212306
#> 26 0.1844614
#> 27 0.2676938
#> 28 0.2172875
#> 29 0.2716288
#> 30 0.2572654
#> 31 0.1981409
#> 32 0.2329732
#> 33 0.2081146
#> 34 0.2005786
#> 35 0.2173499
#> 36 0.2234450
#> 37 0.2276119
#> 38 0.2468265
#> 39 0.2399425
#> 40 0.2424821
#> 41 0.2046557
#> 42 0.2030039
#> 43 0.2333011
#> 44 0.2497746
#> 45 0.2383495
#> 46 0.2563619
#> 47 0.2311343
#> 48 0.2172740
#> 49 0.2060890
#> 
#> attr(,"class")
#> [1] "boot_bw" "list"