Recode

Utility function that recodes variables based on user recode specifications. Handles both numeric or factor variables.

recode(var, recodes, afr, anr = TRUE, levels)

Arguments

var
Variable to recode

recodes
Character string of recode specifications:

Recode specifications in a character string separated by semicolons of the form input=output as in: "1=1;2=1;3:6=2;else=NA"

If an input value satisfies more than one specification, then the first (reading from left to right) is applied

If no specification is satisfied, then the input value is carried over to the result unchanged

NA is allowed on both input and output

The following recode specifications are supported:

Specification	Example	Notes

Single values	9=NA

Set of values	c(1,2,5)=1	The left-hand-side is any R function call that returns a vector

seq(1,9,2)='odd'

1:10=1

Range of values	7:9=3	Special values lo and hi may be used

lo:115=1

Other values	else=NA

Character values are quoted as in :

recodes = "c(1,2,5)='sanitary' else='unsanitary'"

The output may be the (scalar) result of a function call as in:

recodes = "999=median(var, na.rm = TRUE)"

Users are advised to carefully check the results of recode() calls with any outputs that are the results of a function call.

The output may be the (scalar) value of a variable as in:

recodes = "999=scalarVariable"

If all of the output values are numeric, and if 'afr' is FALSE, then a numeric result is returned; if var is a factor then (by default) so is the result.

afr
Return a factor. Default is TRUE if var is a factor and is FALSE otherwise

anr
Coerce result to numeric (default is TRUE)

levels
Order of the levels in the returned factor; the default is to use the sort order of the level names.

Value

Recoded variable

Examples

# Recode values from 1 to 9 to varios specifications
var <- sample(x = 1:9, size = 100, replace = TRUE)

# Recode single values
recode(var = var, recodes = "9=NA")

#>   [1]  7 NA  2  7  4  2  1  8  8  3  4  1  4  5  1  3  6  2  7  6  8 NA  7  5  8
#>  [26]  3  3  1 NA  2 NA  5  8  3 NA  3  7  8 NA  8 NA  5  3  5  4  6  6  7  2  1
#>  [51]  5  5  7  3  1  4  4  3 NA  4  7  6  3  8  3 NA  1  5  4  5  6  8  1 NA  5
#>  [76]  5  3  6 NA  3  8  1  4 NA  1  3 NA  6 NA  7  4  8 NA  2  1  6 NA  7  6  1

# Recode set of values
recode(var = var, recodes = "c(1,2,5)=1")

#>   [1] 7 9 1 7 4 1 1 8 8 3 4 1 4 1 1 3 6 1 7 6 8 9 7 1 8 3 3 1 9 1 9 1 8 3 9 3 7
#>  [38] 8 9 8 9 1 3 1 4 6 6 7 1 1 1 1 7 3 1 4 4 3 9 4 7 6 3 8 3 9 1 1 4 1 6 8 1 9
#>  [75] 1 1 3 6 9 3 8 1 4 9 1 3 9 6 9 7 4 8 9 1 1 6 9 7 6 1

# Recode range of values
recode(var = var, recodes = "1:3=1;4:6=2;7:9=3")

#>   [1] 3 3 1 3 2 1 1 3 3 1 2 1 2 2 1 1 2 1 3 2 3 3 3 2 3 1 1 1 3 1 3 2 3 1 3 1 3
#>  [38] 3 3 3 3 2 1 2 2 2 2 3 1 1 2 2 3 1 1 2 2 1 3 2 3 2 1 3 1 3 1 2 2 2 2 3 1 3
#>  [75] 2 2 1 2 3 1 3 1 2 3 1 1 3 2 3 3 2 3 3 1 1 2 3 3 2 1

# Recode other values
recode(var = var, recodes = "c(1,2,5)=1;else=NA")

#>   [1] NA NA  1 NA NA  1  1 NA NA NA NA  1 NA  1  1 NA NA  1 NA NA NA NA NA  1 NA
#>  [26] NA NA  1 NA  1 NA  1 NA NA NA NA NA NA NA NA NA  1 NA  1 NA NA NA NA  1  1
#>  [51]  1  1 NA NA  1 NA NA NA NA NA NA NA NA NA NA NA  1  1 NA  1 NA NA  1 NA  1
#>  [76]  1 NA NA NA NA NA  1 NA NA  1 NA NA NA NA NA NA NA NA  1  1 NA NA NA NA  1